fix: IFS-2903: uri-builder used where ever possible according to review comments
